moved uploader to components
