+++ src/components/v5/common/AvatarUploader/index.ts
@@ -1,0 +1,1 @@
+export { default } from './AvatarUploader';
