+++ src/components/v5/common/Filter/hooks.ts
@@ -57,16 +57,13 @@ export const useFilter = () => {
     [onSaveSelectedFilters],
   );
 
-  const array: boolean[] = [];
   const isContributorTypeSelected = contributorTypes.some(
     ({ id }) => id === selectedChildOption,
   );
-  array.push(isContributorTypeSelected);
 
   const isStatusTypeSelected = statusTypes.some(
     ({ id }) => id === selectedChildOption,
   );
-  array.push(isStatusTypeSelected);
 
   const isTeamTypeSelected = teamTypes.some(
     ({ id }) => id === selectedChildOption,
@@ -87,11 +84,6 @@ export const useFilter = () => {
     isReputationTypeSelected,
     isPermissionsTypeSelected,
   ].filter(Boolean).length;
-
-  // console.log(array);
-  // console.log({isTeamTypeSelected});
-  // console.log({isContributorTypeSelected});
-  // console.log({checkedItems});
 
   return {
     selectedFilters: [...new Set(selectedFilters)],
